Update application contact name back link to be conditional based on number of registrtion materials.
