+++ src/Epr.Reprocessor.Exporter.UI/Controllers/ReprocessingInputsAndOutputsController.cs
@@ -122,15 +122,15 @@ public class ReprocessingInputsAndOutputsController(
             return Redirect(PagePaths.TaskList);
         }
 
-        session.Journey = [PagePaths.PackagingWasteWillReprocess, PagePaths.ApplicationContactName];
-
         var userData = User.GetUserData();
         var organisationPersons = await GetOrganisationPersons(userData);
 
         var viewModel = new ApplicationContactNameViewModel();
         viewModel.MapForView(userData.Id, currentMaterial, organisationPersons);
 
+        SetJourneyForApplicationContactName(session);
         SetBackLink(session, PagePaths.ApplicationContactName);
+
         await SaveSession(session, PagePaths.ApplicationContactName);
 
         return View(nameof(ApplicationContactName), viewModel);
@@ -255,5 +255,19 @@ public class ReprocessingInputsAndOutputsController(
         var organisationDetails = await accountService.GetOrganisationDetailsAsync(organisationId.Value);
 
         return organisationDetails?.Persons.Where(p => p.UserId != userData.Id) ?? [];
+    }
+
+    private void SetJourneyForApplicationContactName(ReprocessorRegistrationSession session)
+    {
+        var materialCount = session.RegistrationApplicationSession.ReprocessingInputsAndOutputs.Materials.Count;
+
+        if (materialCount == 1)
+        {
+            session.Journey = [PagePaths.TaskList, PagePaths.ApplicationContactName];
+        }
+        else
+        {
+            session.Journey = [PagePaths.PackagingWasteWillReprocess, PagePaths.ApplicationContactName];
+        }
     }
 }
